Moved ale.h to Ale.h
