+++ src/Ale.cpp
@@ -1,4 +1,4 @@
-#include "ale.h"
+#include "Ale.h"
 
 #include <nlohmann/json.hpp>
