+++ tests/ctests/IsdTests.cpp
@@ -4,7 +4,7 @@
 
 #include "gtest/gtest.h"
 
-#include "ale.h"
+#include "Ale.h"
 #include "Isd.h"
 #include "Util.h"
